Use of jump host option for SSH connection
By using the -J option for SSH connection, the script can use any number
of hosts as input. If more than one host is provided, the first hosts will
be considered as jump (gateway) hosts. This avoids using a jump host when on
the intranet. This also greatly simplifies the creation of the tunnel.

The only downside is that sshpass does not behave nicely and kill the
tunnel when connecting to a node. A workaround is in place to check the
connection after establishing the tunnel and issue a warning if
connection fails.